# docs/swagger.yaml
basePath: /
...
info:
  contact:
    name: Fern Platform Team
    url: https://github.com/guidewire-oss/fern-platform/issues
  description: Unified test intelligence platform — aggregate, analyze and act on
    test results from any CI/CD pipeline.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/guidewire-oss/fern-platform
  title: Fern Platform API
  version: "1.0"
paths:
...
  /api/v1/projects:
    get:
      description: Returns a paginated list of all accessible projects
      parameters:
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      - description: Page offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a new project (manager or admin only)
      parameters:
      - description: Project details
        in: body
        name: body
        required: true
        schema:
          properties:
            defaultBranch:
              type: string
            description:
              type: string
            name:
              type: string
            projectId:
              type: string
            repository:
              type: string
            settings:
              type: object
            team:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a project
      tags:
      - projects
  /api/v1/projects/{projectId}:
    delete:
      description: Permanently deletes a project and its data (manager or admin only)
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - projects
    get:
      description: Returns a project by its project ID string
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates project metadata (manager or admin only)
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        schema:
          properties:
            defaultBranch:
              type: string
            description:
              type: string
            name:
              type: string
            repository:
              type: string
            settings:
              type: object
            team:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - projects
  /api/v1/projects/{projectId}/activate:
    post:
      description: Marks a project as active so it accepts new test runs
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Activate a project
      tags:
      - projects
...
  /api/v1/projects/by-project-id/{projectId}:
    get:
      description: Explicit lookup by project ID string (avoids numeric ID ambiguity)
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a project by project ID
      tags:
      - projects
...
securityDefinitions:
  BearerAuth:
    description: 'JWT bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...